+++ changes.md
@@ -1,2 +1,1 @@
-* Port to 1.16.3
-* Do not take into account "minimumLeavesAroundRequired" when using shift down mode
+* Add option to cut trees with mixed log blocks