+++ changes.md
@@ -1,1 +1,1 @@
-* Port to 1.16.3
+* Add option to cut trees with mixed log blocks