+++ changes.md
@@ -1,2 +1,1 @@
-* Check world type before scheduling leaf breaking (#23)
-* Do not take into account "minimumLeavesAroundRequired" when using shift down mode
+* Add option to cut trees with mixed log blocks